Added zombies map rotations and configs.
Use exec gametype.cfg in sv_maprotation to execute the corresponding .cfg for the purpose of running a town survival or other similar type of map. This commit includes all the configs for every gametype featured in zombies. Also includes gamesettings_default.cfg for MP and ZM for reference purposes.